split inner class from service to receiver
