+++ app/src/main/java/ch/bfh/mad/eazytime/remoteViews/notification/ScreenActionService.kt
@@ -1,19 +1,14 @@
 package ch.bfh.mad.eazytime.remoteViews.notification
 
-import android.annotation.SuppressLint
 import android.app.Notification
 import android.app.Service
-import android.content.BroadcastReceiver
-import android.content.Context
 import android.content.Intent
-import android.content.IntentFilter
 import android.os.IBinder
 import android.util.Log
 import ch.bfh.mad.eazytime.TAG
 import ch.bfh.mad.eazytime.di.Injector
 import javax.inject.Inject
 
-@SuppressLint
 class ScreenActionService : Service() {
 
     private val unlockReceiver: ScreenActionReceiver = ScreenActionReceiver()
@@ -44,30 +39,5 @@ class ScreenActionService : Service() {
     override fun onDestroy() {
         Log.i(TAG, "Destroy ScreenActionService and unregister ScreenActionReceiver")
         unregisterReceiver(unlockReceiver)
-    }
-
-    inner class ScreenActionReceiver : BroadcastReceiver() {
-        override fun onReceive(context: Context, intent: Intent) {
-
-            val action = intent.action
-
-            when (action) {
-                Intent.ACTION_SCREEN_ON -> screenOnEvent()
-                Intent.ACTION_SCREEN_OFF -> Log.d(TAG, "screen is off...")
-                Intent.ACTION_USER_PRESENT -> Log.d(TAG, "screen is unlock...")
-            }
-        }
-
-        private fun screenOnEvent() {
-            Log.d(TAG, "screen is on...")
-            notificationHandler.createEazyTimeNotification()
-        }
-
-        fun getFilter(): IntentFilter {
-            val filter = IntentFilter()
-            filter.addAction(Intent.ACTION_SCREEN_OFF)
-            filter.addAction(Intent.ACTION_SCREEN_ON)
-            return filter
-        }
     }
 }
